Project/requirements.txt
flask==3.1.1
flask-cors==5.0.1
google-genai==1.14.0
python-dotenv==1.1.0
requests==2.32.3
